# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "UPI Gateway API",
    "description": "API for creating UPI payment links with QR codes, app-specific deeplinks, IP/CIDR-based access control, Zod validation, and webhook-driven status updates.",
    "version": "1.1.0"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
  "paths": {
...
    "/pay/{id}": {
      "get": {
        "summary": "Render payment page",
        "description": "Displays QR code, UPI app deeplinks, and a countdown timer. Only accessible from allowed IPs/CIDRs. Records first click metadata. Polls /status/:id every 5s for redirects.",
        "tags": ["Payment Page"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment HTML page with QR code, deeplinks, and countdown"
          },
          "403": { "description": "Client IP not in allowed list" },
          "404": { "description": "Payment link not found" },
          "410": { "description": "Payment link expired" }
        }
      }
    },
...
  },
...
}